requirements.txt
Flask==0.12.2
gunicorn==19.7.1
psycopg2-binary==2.8.6
python-dotenv==0.15.0